BDCC/Project 1/app/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-vision
Flask
tensorflow==2.8.0
Pillow==9.1.0